Move: [id]에서 [dateId]로 폴더명 수정 #24
