+++ src/app/(routes)/date-detail/[dateId]/page.tsx
