+++ src/components/features/date-detail/place-name.tsx
@@ -1,4 +1,4 @@
-import { Date } from '@/app/(routes)/date-detail/[id]/page';
+import { Date } from '@/app/(routes)/date-detail/[dateId]/page';
 
 const PlaceName = ({ date }: { date: Date }) => {
   return (
