+++ src/components/features/date-detail/written-content.tsx
@@ -1,4 +1,4 @@
-import { Date } from '@/app/(routes)/date-detail/[id]/page';
+import { Date } from '@/app/(routes)/date-detail/[dateId]/page';
 
 const WrittenContent = ({ date }: { date: Date }) => {
   return (
